feat(codegen/JS/byPass): rework reducer
